fix: chainlink interface

+++ src/oracle/ChainlinkEthOracle.sol
@@ -18,7 +18,7 @@ interface IAggegregatorV3 {
         view
         returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
 
-    function decimals() external view returns (uint256);
+    function decimals() external view returns (uint8);
 }
 
 /// @title ChainlinkEthOracle
